The plugin now runs with the latest API, but...
...it seems to be broken. 

I updated the Plugin based from @ZPlayzMCPE and tested it on my Server. I tried it with an old zip Plugin of mine which worked with the "old" zippluginloader and in addition to that with a random Plugin in .zip Format from Poggit.

The ZipPluginLoader gets enabled but the Plugins were not loaded at all. I had already seen that much Code in this update had been removed in ZPlayz' version.  Nearly all Messages when loading the Plugin had been removed. 


So, now I added back the old Code to find out and test if it still works when updating the functions for the new API. But, im afraid nope… Anyways, my Version of this Plugin is now not really cleaned up, but it should not work much worse than other versions. I hope somebody can fix the Plugin, because I am too low experienced